Update instack-setup-host-rhel7 to use Delorean
Changes instack-setup-host-rhel7 to setup the needed repositories to use with a
midstream Delorean instance. Also adds a git clone of the needed
tripleo-puppet-elements repository. Although this repository is not initially
needed, the work to install the Undercloud via Puppet and deploy the Overcloud
via Puppet will rely on it, so go ahead and add it. Note that
tripleo-puppet-elements is not yet packaged, which is why need to git clone.

Also deletes the previous no longer relevant instack-setup-delorean script.

Change-Id: Ia0e365b15b4b4620c5a392c2a5e3e13d1683b5ef