Update patch set 3
Patch Set 3:

Yes, the finalise stuff is causing me pain as well. I don't have that figured out yet.

For the unregister, make sure you still source the rhel7rc. There's also an updated version of that later in this patch series. I guess that's kind of co-dependent on this patch, so I may squash that one down into this patch.

Patch-set: 3