Ironic-discoverd: Add benchmark data
This patch is adding support to run the AHC benhcmark as part of the
hardware detection in the ironic-discoverd ramdisk.

Two new binaries dependency were added:

  * fio: Used for storage benchmarking
  * sysbench: Used for memory and cpu benchmarking

The patch also removes the installation of the hardware library from
pip to install it from delorean. However, the hardware library has
an unpackaged dependency (ptyprocess). Until we get that packaged,
we will need to install from pip.

Change-Id: I94fe2ba71205fb996700b2c20af396f75900a04e