Set NeutronPublicInterface to nic1
Updates the instack-deploy-overcloud script to use the 'nic1'
os-net-config naming abstraction instead of hard coding this
as eth0.

If you are deploying to real hardware eth0 most likely won't
exist and you'll get an error.

This also drops the settings from the rc scripts since they
no longer need to override the default.

Change-Id: If478f685f1a07ca92f068240a54d5a3275cafb21