Dynamically build separate docs for each environment
The various supported environments require slightly different
setup steps, and rather than use a lot of "If on baremetal..."
lines in the documentation, we can use the sphinx .. only::
directive to specify that certain steps should only be included
in the documentation for certain environments.  This requires
building separate docs for each environment, but since they can
all be built from the same source that shouldn't be a problem.

Change-Id: Ic21c4f91a08d5ef736997cbd599f129830a3c219