Update patch set 1
Patch Set 1: Code-Review-1

(1 comment)

The setup-host script was meant to be short lived. It was intended as just a means to bootstrap our Converged Installer development efforts. That's the reasoning for the baked in rhos-release.

That being said, we can keep it around if people find it useful. The alternative would be to just document the set of repos that are required and document the steps to set them up. Sort of like the existing RDO quickstart which tells you to enable EPEL by installing the right epel-release, then enable RDO by installing the right rdo-release, etc. That might be the best long term solution vs. trying to maintain a setup script. Which approach do you think might work best long term?

At least for this patch as-is, we definitely need to document that the system should already be registered via subscription-manager. We may also want to document how to use $RUN_RHOS_RELEASE.

Also keep in mind that rhos-release is used in instack-build-images as well (via the rhos-release element), so it's usage isn't completely cleaned up. In that case I believe the environment variable used is just $RHOS_RELEASE, so you may consider using the same here.

There's an inline comment as well about an additional repo needed.

This will also need a corresponding change in CI to set $RUN_RHOS_RELEASE. The reason that we don't see a failed subscription-manager command in CI is b/c there is an issue where CI is not using the modified instack-setup-host-rhel7 script from the gerrit review. Attila is going to fix that Monday. Then, CI would be failing on the subscription-manager call here unless $RUN_RHOS_RELEASE is set.

Patch-set: 1
Reviewer: James Slagle <1001877@d5d70762-12d0-45a1-890d-524b12d3f735>
Label: Code-Review=-1