set +x before outputting user messages
We have set -x in a lot of our scripts, which can make output
intended for users a bit difficult to parse because it gets mixed
in with the tracing output.  If we disable tracing before echoing
those messages it cleans up our output a bit, and since those are
always the last steps it shouldn't ever hide any subsequent
operations.

Change-Id: I3189ccfb495371d74801b86f02585c1afdbbef5a