Update patch set 7
Patch Set 7:

this needed to be rebased if you look in the CI failure log.

i've done that in the latest patchset.

However, there are other issues with this as well. First is that the tuskar-ui element expects to do all this:
https://github.com/openstack/tripleo-image-elements/blob/master/elements/tuskar-ui/os-refresh-config/post-configure.d/101-tuskar-ui

Most of which is not going to work. We don't have images for all 4 roles. We also don't need to do all of that sed'ing on the local_settings file now that the undercloud is puppet based.

It might actually be easier to just add the needed packages to be installed to here:
https://github.com/rdo-management/instack-undercloud/blob/master/elements/undercloud-install/package-installs.yaml

and then not use the tuskar-ui element at all.

Patch-set: 7
Label: Code-Review=0