Update patch set 1
Patch Set 1:

Yeah, pretty sure that's why.  If we want to leave this separate I can fix it, but I was thinking maybe we just merge this back into Imre's change so it can all go in together.  Maybe easier than trying to rebase this when he updates the discovery instructions, although I guess if that's the only change it wouldn't be hard.

Patch-set: 1
Label: Code-Review=0