Use a single overcloud-full image.
This moves us to using a single overcloud-full image for all of the different
overcloud roles. Puppet will handle enabling and starting only the applicable
services for the role for which the image is deployed.

This was recorded as an outcome in:
https://trello.com/c/RlZDnl8a

This will greatly simplify the number of images we need to build and maintain
going foward.

Change-Id: I495961847c1745d59de4bfac5967bd6bcf2e13c6