Make running the benchmarks during discovery optional
Not every use case will require the benchmark data, and it can add
a significant amount of time to the discovery process. This patch
adds a RUNBENCH kernel command line option to the discovery
ramdisk. If it is set to 1, the benchmarks will run. The default
is to not run them.

Change-Id: Ib64ff5f1d34de6a2ab6ed3f3a9ee478f697b87c6