Get ready for Ironic state machine changes
After rebase of ironicclient discoverd will no longer work with
newly created nodes, as they will be in AVAILABLE state, where
discovery is not allowed. Move nodes to MANAGEABLE state right
after creating, move them back to AVAILABLE just before deploy.

This change fixes IRONIC_API_VERSION in scripts to be 1.6,
so that we're not broken again on sudden changes.

Change-Id: I737fcc04a72eb17c8f7cbc65f6b29917b1706b9d