Update patch set 3
Patch Set 3: Code-Review+2

Should probably have a readme for the new element, but a bunch of our current elements don't and this one is trivial enough that I'm not inclined to block it for that.

Worked fine for me locally, so this should be good to go.  It does need to pass CI, which hasn't run yet because the change has no topic.  I'll take care of that.

Patch-set: 3
Label: Code-Review=+2