Update patch set 4
Patch Set 4:

I don't think that doing it in discover_nodes and all the ready state functions is a right way to go. Things can easily become messy that way.

Actually this patch does the right thing, if we have problems with some workflow we should fix it, I guess. Could you elaborate on "rediscover workflow" as it's the first time I'm hearing about it, and it does not make sense to me with my upstream hat on.

The node state flow must be:
1. Node is enrolled
2. Node is in MANAGEABLE state
3. Discovery, ready state, etc happens
4. Node is in AVAILABLE state
5. Deploy

If hardware is upgraded later, then:
6. Undeploy
7. Put in maintenance mode and MANAGEABLE
8. Upgrade hardware
9. Remove maintenance mode
10. Rediscover
11. Put to AVAILABLE
12. Deploy again

What are we doing wrong?

Patch-set: 4
Reviewer: Dmitry Tantsur <1003460@d5d70762-12d0-45a1-890d-524b12d3f735>
Label: Code-Review=0