Update patch set 4
Patch Set 4:

Okay, fair enough.  The problem is that in practice people sometimes don't run discovery just once.  They might need to run it multiple times because their specs didn't match as expected (this is actually a tangential problem that exists today), or they just decide to change something after deployment.  At the very least we need to document how to do that with this new workflow, which is currently in step 3 of https://repos.fedorapeople.org/repos/openstack-m/instack-undercloud/html/deploy-overcloud.html#redeploying-the-overcloud

I also think we need to put something in the discovery section that explains how to re-run discovery if it doesn't work right the first time.  As I mentioned above, that's separate from this change, and maybe can piggy-back on https://review.gerrithub.io/#/c/229536  We could just link to an FAQ about re-running discovery so we don't clutter up the overcloud docs.

Anyway, for this change we need to update the existing docs to reflect the changes, and I think we need some other changes related to this workflow to explain how different scenarios should be done.

Patch-set: 4