Update patch set 4
Patch Set 4:

Dmity, I think the important distinction is that your steps 3 and 4 in the first set need to be able to happen multiple times. That might be why it's easier to just put the node state setting into discover_nodes.

I understand why you might not want to do that. So, we can go the documentation route as well that tells people how to re-do discovery in such a way that it will work. Keep in mind that we aren't building out instack-ironic-deployment to be an end user CLI or tool with the right fool-proof workflow. We really just want it so that if you run --discover-nodes, it works, without having to be aware of the valid states for the Ironic state machine.

The important bit is that if you follow the Sphinx documentation, you can exercise the features we're trying to let people use. Run discovery multiple times is an important part of that (for the reasons Ben mentioned).

Patch-set: 4
Reviewer: James Slagle <1001877@d5d70762-12d0-45a1-890d-524b12d3f735>
Label: Verified=0