Support Undercloud Log Aggregation
This commit introduces support for undercloud log aggregation,
which is composed of a fluentd log aggregator and an elastic search
instance.

Configuration of fluentd and elasticsearch must be enabled by
setting `ENABLE_LOG_AGGREGATION=true` in 'instack.answers'.

This commit requires the corresponding commit in
tripleo-puppet-elements, in order to have the required
puppet module installed.

Depends-On: Ib739de0ca70b8f85c0b4fb3c9ea57fd9c2b0fa61@review.openstack.org
Change-Id: I7eaf3cba097fb61ab5932d7e8313089417f346b1