Update patch set 5
Patch Set 5: Code-Review-1

(4 comments)

Mostly looks good, -1 is for wrong heading in troubleshooting-nodes.rst

I'd like us to also consider terms using for "discovery", because the way we use word "discovery" is confusing (and not accepted upstream). But I leave it up to you whether to fix it now.

Patch-set: 5
Label: Code-Review=-1