Update patch set 3
Patch Set 3:

Well basically I would love to have Introduction to be the landing page in the future. But first of all I need to figure out how, and second I didn't want to block this improvements. We can iterate on it, so I kept current index as it is now.

Patch-set: 3
Label: Code-Review=0