Fixed ImportError: No module named six on Centos
When running instack-install-undercloud on CentOS 7.1, the following
error occurs:

"ImportError: No module named six on Centos"
from /tmp/tmpcPpZxn/extra-data.d/10-merge-svc-map-files", line 18

This is because the python-six  is not included in a base install
of CentOS 7.1.  This changes adds installation of this RPM to
instack-setup-host.

This changes fixes BZ: 1212862

Change-Id: I480519494fd98f6ee252b9ad5032f4f6926adb03