Support Overcloud Log Aggregation
This commit updates the 'instack-build-images' script to include
the 'overcloud-logging' element, which will install the packages
required for log aggregation in the overcloud image.

Additionally, this commit adds a new option to 'instack-deploy-overcloud',
`--log-aggregation-server HOSTNAME_OR_IP`, which enables the actual
configuration of the log aggregation agents on the overcloud.

This commit requires the corresponding commits to tripleo-puppet-elements
and tripleo-heat-templates to function properly.

Depends-On: Ic0af2ca635e7dad27945fcbee2df8afe2c72a83c@review.openstack.org
Change-Id: I2b3c7b65560753fad7290c25cd3fe4a4b61d6601