Support Overcloud Log Aggregation
This commit updates the 'instack-build-images' script to include
the 'overcloud-logging' element, which will install the packages
required for log aggregation in the overcloud image.

This commit requires the corresponding commit to tripleo-puppet-elements
to function properly.

Additionally, this commit adds a new option to 'instack-deploy-overcloud',
`--log-aggregation-server HOSTNAME_OR_IP`, which enables the actual
configuration of the log aggregation agents on the overcloud.