Update patch set 6
Patch Set 6: Code-Review-2

> (2 comments)
 > 
 > > > I fixed the wording to cover your concerns. So...
 > > >
 > > > What do you mean by "no support"? If we provide how the virtual
 > > > environment should look like (very similar way we are doing
 > that
 > > > with baremetal), where is the problem?
 > >
 > > *If* we provide how the virtual environment should look, there
 > > wouldn't be a problem! You're the one who said you couldn't
 > afford
 > > to do that as part of this change. Therefore, I don't think this
 > > change should go in as-is.
 > 
 > Therefor I've added this part:
 > .. note::
 > These virtual machines will be automatically created in the
 > following section
 > named :ref:`preparing_virtual_environment`.

You've reversed the order of what we said we were going to setup, and what's needed or actually gets setup. This changes the meaning and makes it open to interpretation about if you even need to do the 2nd part, especially since you've represented the part about running the Setup as a Note.

My -1 keeps getting lost on this point, so bumping to -2. I feel pretty strongly about this. Especially since you're changing the expectation around how to get setup, and effectively leaving it to others to then support and help folks out when they get stuck.

Patch-set: 6
Label: Code-Review=-2