Quintupleo setup documentation
Adds setup docs for quintupleo environment as an alternative to
virt and baremetal

I still haven't got an overcloud up, but it would be useful for others
to try now and see how far they get.

There are admonitions describing unresolved quintupleo issues but it
should now be possible to at least have completed node registration.

Change-Id: Ic5d059320458a8fa275d417a9f4bd8f410f4754d