Update patch set 9
Patch Set 9: Code-Review-1

(3 comments)

A few minor comments inline.  -1 partially because this needs to wait for CI to be updated, and it's going to conflict with the general troubleshooting change that I think is ready to merge.  This should probably be rebased on top of that change.

If there's a proposed CI change it would be good to add that as a Depends-On in the commit message too.

Patch-set: 9
Reviewer: Ben Nemec <1003493@d5d70762-12d0-45a1-890d-524b12d3f735>
Label: Code-Review=-1