Update patch set 1
Patch Set 1: Code-Review-1

Adding new and new admonitions is unsustainable. Not saying that you this patch is the one which is causing the problem, but with this patch it is already becoming unbaerable, it is completely breaking the user flow. Majority of people will not deploy QuintupleO and these admonitions are huge distraction. We need to solve for this use-cases.

You are mentioning that some commands are failing due to bugs. In that case why to add QuintupleO docs until it is working?

Patch-set: 1
Reviewer: Jaromír Coufal <1003879@d5d70762-12d0-45a1-890d-524b12d3f735>
Label: Code-Review=-1