Matching is seperated from introspection.
This changes the documentation to reflect the seperation of
matching and introspection.

We can also revert the change to the /etc/edeploy/ configuration
files as a default profile is no longer needed for the basic flow.

This further adds some default config for the ready-state workflow,
so that the advanced flow works.

Change-Id: If88fc8cf69aa00f81098398e7ac643ff31d57959