Update patch set 10
Patch Set 10: Code-Review-1

(9 comments)

thanks for addressing my flow concerns. this is looking good. just a few nits inline that i think warrant fixing.

and also, the link to the arch doc talks about the deploy/workload cloud again. I'd really like to avoid that if we can until we settle on adopting those terms or not.

Patch-set: 10
Label: Code-Review=-1