Update patch set 7
Patch Set 7:

Somewhere in the installation process, we need to make sure that the set of services being tested by tempest is the same as the services that are being deployed. The tempest config script will use discovery on the catalog to determine which tests to run leaving the failure to detect the bug if an expected service somehow is not in the catalog. The solution is for the installer to write a partial tempest conf file with the [service-available] section provided. This file can then be passed in to the config script using the --deployer-input argument. The contents of of this file are treated similarly to the command line  argument. This file should also really contain any other values that are known only to the installer such as the names of some backend parameters that are used in the cinder tests and must match the values in cinder.conf.

Patch-set: 7
Reviewer: David Kranz <1002284@d5d70762-12d0-45a1-890d-524b12d3f735>
Label: Code-Review=0