Create and set permissions for the right rabbitmq user
In the puppet manifest, we were creating rabbit users with the service
names (nova, glance, etc), and verifying those permissions. However,
those weren't the actual rabbitmq users we were using. The user we
actually use is set by $UNDERCLOUD_RABBIT_USERNAME, which defaults to
"guest" and is written to the hiera file. This updates the puppet
manifest to read the rabbit username from hiera when creating the user
and validating it's permissions.

Change-Id: I70ec81d2324bbba34ee74a6e3c0d126259ade36e