Update patch set 1
Patch Set 1:

Then we shouldn't have these functions at all in our public source code. Which means that we have to maintain an internal fork of instack-undercloud (which sounds insane, but it wasn't my who asked for that). Then we can build our own HTML docs (because Mojo is never a good answer).

Or just leave it alone and stop pretending that we don't have anything internal...

Patch-set: 1