Update patch set 1
Patch Set 1: Code-Review-1

(14 comments)

First, I think we need to separate patches based on Chapters.

Second, we need to be more advanced in explanation why is that part needed in RDO-Manager, if it is required/optional, how to do it and how to apply this to my (user's) environment. If something goes wrong where to go. + links to upstream documentation about the feature in general.

Patch-set: 1
Label: Code-Review=-1