Update delorean RDO-Manager URL
With the release of Kilo upstream, we now need to be basing our
changes on the stable/kilo branch upstream. The RDO-Manager
delorean instance has been updated to build multiple releases
simultaneously. In order to do that, the URL has changed. This
patch updates instack-undercloud to use the new URL.

Change-Id: I21c0fefa5e83f8e05abe0893d3a98839fa706486