Add proxy switch to curl
When building undercloud image, we rely on copied env variables and
default configuration of sudoers.d By default, sudo is not allowed
to keep env variables like proxies.
This is a problem, when user uses non-transparent proxy, because
every time when chroot is executed, we lose proxy environment
variables. It causes problem with downloading repo from net and
building of image cannot be successfully finished.
To bypass this problem, take proxy and pass it to curl.

Change-Id: Ice90916bbf9a3a7b468f4e68cd783f445c806cfc