Update patch set 1
Patch Set 1:

Feel free to reject this as it's very subjective.

I got a bit confused when I saw the instructions that I was installing the unified CLI but the RPM was a plugin. I get that the dependencies picked it up, but my first reaction was that something was mislabeled. It's possible (likely) that normal users won't be as concerned with potential bugs in the docs as I am right now as we come up on the release, so I won't make a huge argument if we want to leave it as is.

Patch-set: 1
Label: Code-Review=0