Update patch set 3
Patch Set 3: Code-Review-1

Is there any reason we couldn't just pass -E to sudo to preserve the environment variables?  If that works I'd prefer it to munging the command with regexes like this.

pep8 failed because this patch is based on an old branch where pep8 wasn't working yet.  It does make me realize that logs are no longer being uploaded for failed unit test runs though. :-/

Patch-set: 3
Label: Code-Review=-1