Update patch set 1
Patch Set 1:

(1 comment)

Originally, the cmdb wasn't designed for ready-state but we started using it because that was the easiest move. You can put anything there really, it's the responsibility of Ironic to use that data in our use-case. (...and for now, those 2 keys are the only supported ones.)

The original documentation of cmdb:
https://github.com/enovance/edeploy/blob/master/docs/eDeployUserGuide.rst#using-the-configuration-management-database-cmdb 

If you think it would make sense to add this URL, I will add that but eg. ready-state definition doesn't work inside the generate calls.

Patch-set: 1
Label: Code-Review=0