Force nova config drive to True in puppet data
Currently if you follow the advanced discovery/introspection flow
you find that the dicovery data is not injected into the deployed
nodes via the config-drive as expected.

This change configures the undercloud nova to always include the
configdrive data such that it's available for deploy-time scripts
to consume, e.g to create the mappings for os-net-config or
configure filesystems based on knowledge of disk performance/type.

Change-Id: I8dac26dab4eeccb3e8c878d51ce9d3944b4f7675