Update patch set 1
Patch Set 1:

> Two issues here: 
> 1. new architecture information duplicates what you've added in introduction. what about just extending introduction? 

Yes, that's actually what I started out doing (see first patch I uploaded for the introduction), but it ended up getting pretty detailed for an "introduction" page.

The intention here wasn't to duplicate, but to reiterate the top-level of the architecture with a little more detail and specifics than the (not very detailed) introduction.

My idea was this sets the scene for a series of sections getting increasingly into architectural details on each subsystem.

I'm happy to refactor this into one-big introduction page though if that's what folks prefer :)

>2. architecture page is very hard to find. that's not your fault, but should be noted.

Yeah, I played with promoting it to a top-level section, e.g so we have "Introduction" with some sections giving and overview, and an "Architecture" section with a piece-by-piece decomposition and detailed discussion of the design.

Maybe we should do that?

I'm happy to hold off merging this for now and I'll push some follow-up patches then we can see where the information sits most logically.

Patch-set: 1