Improve discovery image build time/size
This does a couple of things that reduce the time taken to build
the discovery image, and also the size of the final image.  First,
it generates a list of the files that need to installed, then uses
dracut's inst_multiple function to install them rather than using
inst on each individual file.  This seems to provide a small speed
improvement.

The major improvement is omitting pyc and pyo files from the image.
These are simply a Python optimization that is not required, and
we likely lose more time transferring the extra files to the target
system than we gain in improved module load speed.

On a reasonably fast system, the two techniques cut about 5 minutes
and 20 MB from the image build.

Change-Id: I75318a09689fbce59a61d05d427b6acb5f251064