Update patch set 4
Patch Set 4: Verified+1

This has actually had a CI pass previously, so going to merge it.  We'll have to transfer the overcloudrc docs to the ucli, but at least the information will be in the repo.

Patch-set: 4
Reviewer: Ben Nemec <1003493@d5d70762-12d0-45a1-890d-524b12d3f735>
Label: Verified=+1