Use plan name instead of uuid for deployment
The unified cli now supports deploying plans by name, so let's
make use of that.  Having to manually look up the uuid for a basic
deployment is just one more opportunity for mistakes.

If we come up with a use case where there are multiple plans, that
will likely be an advanced deployment topic so it doesn't belong in
basic_deployment anyway.

Change-Id: Id48e039aea6e5f291ee5804eb1a675a6b61c7087