Additions to Network Documentation for External bridge
This change adds documentation for using floating IP networks. When the
floating IP network will be on a tagged VLAN, the
NeutronExternalNetworkBridge should be set to "''", which will result
in the Neutron configuration of floating IP networks being attached to
br-int rather than br-ex. If the floating IP network will be on
the native VLAN, then the setting should be changed to "br-ex".

This change also makes a minor update to the ServiceNetMap, to reflect
changes that were made to support mapping hostnames onto a network. The
IP on the stated network will be used to resolve the hostname.

Change-Id: Ic3727a13a88a195a65a3c189513b6ae0656a1138