Fix Horizon undercloud UI login
There were 2 issues causing problems logging into the undercloud UI.

First, when SSL was enabled, Horizon was failing to login with the self
signed cert because it's configured to use the publicURL endpoints. This
reconfigures Horizon to use the internalURL endpoints which are always
non-SSL.

Secondly, grant the admin user the admin role in the service project
since no project is specified when logging to the tuskar ui, and the
admin role is always required and we have no way to control what project
will be looked at if one isn't specified. See the bugzilla for more
details.

Change-Id: I688b40601359641e8963e6655e1ecf86dfa90792
Resolves: rhbz#1245169