Update patch set 1
Patch Set 1: Code-Review-1

According to this [1] it is possible to specify default project for the user when is being created. This way the default project will be used when user is logged in instead of choosing it semi-randomly when no default project is specified.

So I think better solution is to specify admin project as default project for admin user when he is created. [2]

[1] https://github.com/openstack/django_openstack_auth/blob/master/openstack_auth/forms.py#L41-L46
[2] http://docs.openstack.org/user-guide-admin/manage_projects_users_and_roles.html#create-a-user

Patch-set: 1