Updates docs for for recreation of deployment plan and roles
Adds a section in Advanced Deployment to document how to delete and
recreate the deployment plan and roles from scratch. This includes
a way to specify an alternative tripleo heat templates directory
from which roles are created.

A related bug is at
https://bugzilla.redhat.com/show_bug.cgi?id=1229493

In time we can change the procedure documented here to use the
--all flags to tuskar load/delete roles as tracked in that BZ
and avoid using the tripleo image element scripts directly.

Change-Id: I099dde62909f576b32039d594a397a59d2b11b4f