Configure nova for use with Ironic
This patch updates the nova.conf to include several settings
that are desirable for use when using the Ironic compute
driver. The following new settings are configured with this
patch:

 scheduler_host_manager=nova.scheduler.ironic_host_manager.IronicHostManager
 reserved_host_memory_mb=0
 compute_manager=ironic.nova.compute.manager.ClusteredComputeManager
 scheduler_tracks_instance_changes=False

Without these settings I'm unable to create an overcloud with
settings that worked on the TripleO seed. Specifically, I'm using
failures caused by a variety of scheduler filters which are incorrectly
matching memory and disk usage for my baremetal machine.

These settings bring our instack undercloud inline with the
upstream recommended Ironic settings for nova.conf here:

 http://docs.openstack.org/developer/ironic/deploy/install-guide.html

Change-Id: I52e16b95418771750a89d150da44ac77ccffa6d5