Support relative certificate paths
If the service certificate specified in undercloud.conf was a
relative path, the install would fail because the value was written
to haproxy.cfg verbatim.  This changes the behavior to look up the
absolute path so the value written to the config will be valid
even though haproxy will not be running in the same working
directory as the installer.

Change-Id: Ida979428f289d59bb0be356cc0c61ab73f1a67c2