Updates to network documentation to accompany y1
This change contains fixes to the network documentation as well as
additions of the creation of floating IP and provider networks. The
bonding information was also significantly updated. All references
to --plan overcloud were replaced with --templates. A bug in the docs
referenced an incorrect way to define the NeutronExternalNetworkBridge,
this is fixed now.

Change-Id: Id2f8464e4ee500d65ea27c1816f8ed94906fdf36