Update patch set 2
Patch Set 2:

There should really be a chapter in our docs about using tempest but the doc team has not gotten to that. The tempest README.rpm is not specific to usage with director and is not a great target to be pointed to. I agree this should be in one place but until we can identify where that is, putting a few words here is the least of evils I think.

Patch-set: 2
Reviewer: David Kranz <1002284@d5d70762-12d0-45a1-890d-524b12d3f735>
Label: Code-Review=0